pyproject.toml
[project]
name = "nyt-pieces"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "altair>=5.0.0",
    "pandas>=2.0.0",
    "vegafusion>=1.0.0",
    "vl-convert-python>=1.0.0",
    "pyarrow>=14.0.0",
    "scikit-learn>=1.3.0",
    "beautifulsoup4>=4.14.3",
    "numba>=0.59.0",
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.41",
    "bertopic>=0.17.0",
    "networkx>=3.0",
    "spacy>=3.8.11",
    "en-core-web-sm",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
